refactor(cli): make agent spawn/detect helpers CGO-free
The upcoming `lk session` daemon ships in the default (CGO-free) `lk`
binary, but the logic to detect which agent to run, build its console
argv, and spawn/signal the subprocess lived behind the `console` build
tag — which pulls in PortAudio/APM via CGO. None of that logic is
actually CGO; it was tagged only because it sat next to the console
command. This frees the pure-Go pieces without relocating any audio code.

- Untag simulate_subprocess.go (AgentProcess + spawn/entrypoint helpers).
- Retag proc_unix.go / proc_windows.go from `console && <os>` to `<os>`,
  since the now-untagged spawner calls their signal/attr helpers. Add
  setDetachedProcAttr (Setsid / CREATE_NEW_PROCESS_GROUP) for the daemon.
- Extract detectProject + buildConsoleArgs into a new untagged
  agent_utils.go, shared by `lk agent console` and `lk session`.

No behavior change: every moved function is identical and nothing is
rewired. Both build modes stay green (CGO-free default and -tags console).
setDetachedProcAttr is unused until the session daemon lands.
