+++ cmd/lk/agent_run.go
@@ -22,13 +22,10 @@ import (
 	"io"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"sync"
 	"syscall"
 
 	"github.com/urfave/cli/v3"
-
-	"github.com/livekit/livekit-cli/v2/pkg/agentfs"
 )
 
 func init() {
@@ -107,45 +104,6 @@ func resolveCredentials(cmd *cli.Command, loadOpts ...loadOption) ([]string, err
 		args = append(args, "--api-secret", apiSecret)
 	}
 	return args, nil
-}
-
-func detectProject(cmd *cli.Command) (string, agentfs.ProjectType, string, error) {
-	explicit := cmd.Args().First()
-
-	detectFrom := "."
-	if explicit != "" {
-		absPath, err := filepath.Abs(explicit)
-		if err != nil {
-			return "", "", "", err
-		}
-		if _, err := os.Stat(absPath); err != nil {
-			return "", "", "", fmt.Errorf("entrypoint file not found: %s", explicit)
-		}
-		detectFrom = filepath.Dir(absPath)
-	}
-
-	projectDir, projectType, err := agentfs.DetectProjectRoot(detectFrom)
-	if err != nil {
-		return "", "", "", noAgentError()
-	}
-	if !projectType.IsPython() {
-		return "", "", "", fmt.Errorf("currently only supports Python agents (detected: %s)", projectType)
-	}
-
-	if explicit != "" {
-		absPath, _ := filepath.Abs(explicit)
-		rel, err := filepath.Rel(projectDir, absPath)
-		if err != nil {
-			return "", "", "", fmt.Errorf("entrypoint %s is outside project root %s", explicit, projectDir)
-		}
-		return projectDir, projectType, rel, nil
-	}
-
-	entrypoint, err := findEntrypoint(projectDir, "", projectType)
-	if err != nil {
-		return "", "", "", err
-	}
-	return projectDir, projectType, entrypoint, nil
 }
 
 func buildCLIArgs(subcmd string, cmd *cli.Command, loadOpts ...loadOption) ([]string, error) {
