+++ cmd/lk/proc_windows.go
@@ -1,4 +1,4 @@
-//go:build console && windows
+//go:build windows
 
 package main
 
@@ -6,9 +6,16 @@ import (
 	"os"
 	"os/exec"
 	"strconv"
+	"syscall"
 )
 
 func setProcAttr(_ *exec.Cmd) {}
+
+// setDetachedProcAttr starts the daemon in a new process group so it is not
+// killed when the parent CLI invocation exits.
+func setDetachedProcAttr(cmd *exec.Cmd) {
+	cmd.SysProcAttr = &syscall.SysProcAttr{CreationFlags: syscall.CREATE_NEW_PROCESS_GROUP}
+}
 
 func (ap *AgentProcess) sendInterrupt() {
 	if ap.cmd.Process != nil {
